+++ src/db/schema/mod.rs
@@ -18,6 +18,7 @@ mod v25_to_v26;
 mod v26_to_v27;
 mod v27_to_v28;
 mod v28_to_v29;
+mod v29_to_v30;
 mod v2_to_v3;
 mod v3_to_v4;
 mod v4_to_v5;
@@ -55,6 +56,7 @@ use v25_to_v26::UPGRADE_25_TO_26_SQL;
 use v26_to_v27::UPGRADE_26_TO_27_PRE_SQL;
 use v27_to_v28::migrate_v27_to_v28;
 use v28_to_v29::{migrate_v28_to_v29, UPGRADE_28_TO_29_PRE_SQL};
+use v29_to_v30::{migrate_v29_to_v30, UPGRADE_29_TO_30_PRE_SQL};
 use v2_to_v3::UPGRADE_2_TO_3_SQL;
 use v3_to_v4::UPGRADE_3_TO_4_SQL;
 use v4_to_v5::UPGRADE_4_TO_5_SQL;
@@ -166,5 +168,10 @@ pub const MIGRATION_27_TO_28: Migration = Migration::standalone(migrate_v27_to_v
 pub const MIGRATION_28_TO_29: Migration = Migration::Transacted {
     pre_sql: Some(UPGRADE_28_TO_29_PRE_SQL),
     code_fn: Some(migrate_v28_to_v29),
+    post_sql: None,
+};
+pub const MIGRATION_29_TO_30: Migration = Migration::Transacted {
+    pre_sql: Some(UPGRADE_29_TO_30_PRE_SQL),
+    code_fn: Some(migrate_v29_to_v30),
     post_sql: None,
 };